requirements.txt
anytree==2.8.0
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-admin-csvexport==2.0
django-admin-interface==0.24.2
django-admin-tools==0.9.2
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-import-export==3.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.4.0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
xlrd==2.0.1
xlwt==1.3.0
